+++ MemfaultPackages/bort/src/main/java/com/memfault/bort/settings/Settings.kt
@@ -53,6 +53,8 @@ interface LogcatSettings {
     val collectionInterval: Duration
     val commandTimeout: Duration
     val filterSpecs: List<LogcatFilterSpec>
+    val kernelOopsDataSourceEnabled: Boolean
+    val kernelOopsRateLimitingSettings: RateLimitingSettings
 }
 
 interface FileUploadHoldingAreaSettings {