+++ MemfaultPackages/bort/src/main/java/com/memfault/bort/tokenbucket/TokenBucketFactory.kt
@@ -9,13 +9,19 @@ abstract class TokenBucketFactory(
     val defaultCapacity: Int,
     val defaultPeriod: Duration,
 ) {
-    abstract fun create(count: Int, capacity: Int, period: Duration): TokenBucket
+    abstract fun create(count: Int, capacity: Int, period: Duration, periodStartElapsedRealtime: Duration?): TokenBucket
 
-    fun create(count: Int? = null, capacity: Int? = null, period: Duration? = null): TokenBucket =
+    fun create(
+        count: Int? = null,
+        capacity: Int? = null,
+        period: Duration? = null,
+        periodStartElapsedRealtime: Duration? = null
+    ): TokenBucket =
         create(
             count = count ?: defaultCapacity,
             capacity = capacity ?: defaultCapacity,
             period = period ?: defaultPeriod,
+            periodStartElapsedRealtime = periodStartElapsedRealtime,
         )
 }
 
@@ -25,12 +31,17 @@ class RealTokenBucketFactory(
     defaultCapacity: Int,
     defaultPeriod: Duration
 ) : TokenBucketFactory(defaultCapacity, defaultPeriod) {
-    override fun create(count: Int, capacity: Int, period: Duration): TokenBucket =
+    override fun create(
+        count: Int,
+        capacity: Int,
+        period: Duration,
+        periodStartElapsedRealtime: Duration?
+    ): TokenBucket =
         TokenBucket(
             capacity = capacity,
             period = period,
             _count = count,
-            _periodStartElapsedRealtime = realElapsedRealtime(),
+            _periodStartElapsedRealtime = periodStartElapsedRealtime ?: realElapsedRealtime(),
             elapsedRealtime = ::realElapsedRealtime,
         )